% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_file}
\alias{cache_file}
\title{Cache files}
\usage{
cache_file(obj, file_name)
}
\arguments{
\item{obj}{The object to cache. Should be a simple features object, data.frame,
tibble, or matrix.}

\item{file_name}{String detailing desired file_name. If file type is included,
should be one of ".csv" or ".shp"}
}
\value{
None.
}
\description{
Take a simple features, dataframe, or tibble object and save it to a cache
directory.
}
\seealso{
\link{clear_cache}
}
