% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{cache_geometries}
\alias{cache_geometries}
\title{Cache all OEPS geometries.}
\usage{
cache_geometries(scales = "all")
}
\arguments{
\item{scales}{Or list of strings specifying which geometries to cache.
Must be one of "state", "tract", "county", "zcta" or "all". Defaults to "all."}
}
\description{
Downloads and caches all OEPS geometries, overwriting any previously cached
geometries.
}
