% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_oeps.R
\name{cache_oeps_tables}
\alias{cache_oeps_tables}
\title{Cache OEPS data tables.}
\usage{
cache_oeps_tables()
}
\value{
None
}
\description{
Retrieve most recent version of all OEPS attribute data tables and save them
to cache. To cache geometries, see \code{\link[=cache_geometries]{cache_geometries()}}. To clear cached
data, see \code{\link[=clear_cache]{clear_cache()}}
}
