% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cache}
\usage{
clear_cache()
}
\value{
0 on a successful cache deletion and 1 on a failure. Deleting a
nonexistent cache counts as a success.
}
\description{
Remove all files from the oepsData cache, and delete the cache file itself.
}
