% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_by_county}
\alias{filter_by_county}
\title{Filter by county}
\usage{
filter_by_county(df, counties)
}
\arguments{
\item{df}{Dataframe to filter. Must contain HEROP_ID.}

\item{counties}{String or vector of strings detailing which counties to filter to.}
}
\value{
Dataframe containing only observations in the requested counties.
}
\description{
Filters a given dataframe to only include observations from a given
county. Dataframe must possess a HEROP_ID column.
}
