% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_formatting.R
\name{filter_by_geography}
\alias{filter_by_geography}
\title{Filter by geographies}
\usage{
filter_by_geography(df, states, counties)
}
\arguments{
\item{df}{Dataframe to filter.}

\item{states}{String or vector of strings detailing which states to filter on.
Can be the states name, abbreviation, or GEOID. Optional. If null, counties must
either not be specified or must be GEOIDS. Defaults NULL.}

\item{counties}{String or vector of strings detailing which counties to filter on.
Can be county names, abbreviations, or GEOIDS. Optional. If multiple counties are
passed, counties must either be from the same state or be GEOIDS.}
}
\value{
Dataframe containing only observations which satisfy the filter.
}
\description{
Filters a given dataframe to only include entires with specified state and
counties. If no state or counties are specififed, does nothing.
}
