% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_by_state}
\alias{filter_by_state}
\title{Filter by state}
\usage{
filter_by_state(df, states)
}
\arguments{
\item{df}{Dataframe to filter. Must contain HEROP_ID.}

\item{states}{String or vector of strings detailing which states to filter by.}
}
\value{
Dataframe containing only observations which occurred in a given state.
}
\description{
Filters a given dataframe to only include entries from a given state.
Dataframe must possess a HEROP_ID column.
}
