% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_oeps.R
\name{filter_by_themes}
\alias{filter_by_themes}
\title{Filter by theme}
\usage{
filter_by_themes(attribute_data, themes)
}
\arguments{
\item{attribute_data}{data.frame containing the attribute data to filter on.}

\item{themes}{List of themes to filter out.}
}
\value{
Filtered data.frame.
}
\description{
Takes user defined list of themes and references data dictionary to
filter down to only variables within that list as well as geography variables.
}
