% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_oeps.R
\name{get_attribute_table}
\alias{get_attribute_table}
\title{Download attribute data}
\usage{
get_attribute_table(scale, year, cache = FALSE)
}
\arguments{
\item{scale}{Same as in \code{load_oeps}.}

\item{year}{Same as in \code{load_oeps}.}

\item{cache}{whether to cache newly retrieved data and/or check for previously
cached data.}
}
\value{
OEPS data table containing data with the requested scale and year.
}
\description{
Reads in attribute data tables from GitHub given the desired scale and year.
}
