% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{get_cached_geometry_name}
\alias{get_cached_geometry_name}
\title{Cached geometry name}
\usage{
get_cached_geometry_name(scale, cb = FALSE, year = 2010)
}
\arguments{
\item{scale}{Character specifying scale of the requested geometry. Should be
one of "state", "tract", "zcta", or "county".}

\item{cb}{Boolean specifying whether or not the file is a cartographic boundary
file. Defaults to FALSE.}

\item{year}{Numeric or character specifying the year the geometries correspond
to. Defaults to 2010.}
}
\value{
string containing the file name of geometry satisfying the given
parameters.
}
\description{
Retrieve the file_name -- but not location -- of a cached geometry file.
}
