% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_oeps.R
\name{load_oeps}
\alias{load_oeps}
\title{Load OEPS data}
\usage{
load_oeps(scale, year, themes = "All", geometry = FALSE)
}
\arguments{
\item{scale}{String specifying which geographic scale to pull variables for.
Must be one of "tract", "zcta", "county", or "state".}

\item{year}{String or integer specifying which year to pull variables for.
Must be one of "1980", "1990", "2000", "2010", or "Latest".}

\item{themes}{String or list of strings specifiying which Opioid Risk Environment
themes to pull variables for. Valid themes are "Geography", "Social", "Environment",
"Economic", "Outcome", "Policy", "Composite", or "All". Defaults to "All".}

\item{geometry}{Boolean specifying whether to pull geometries for the dataset
(TRUE) or not (FALSE). Defaults FALSE.}
}
\value{
A tibble containing the requested variables. Always exports merge keys.
}
\description{
\code{load_oeps} packages opioid environment policy scan data with given geographic
scale, year, and themes. Can also pull geometries.
}
\examples{
load_oeps('tract', 2010, geometry=TRUE)
load_oeps('county', 'latest', themes='social')

# themes can also be a list of options
load_oeps('zcta', 'latest', c('composite', 'economic'))

}
\seealso{
\code{\link[=load_oeps_dictionary]{load_oeps_dictionary()}}, which loads a basic data dictionary
showing viable datasets.
}
