% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_oeps_dictionary.R
\name{load_oeps_dictionary}
\alias{load_oeps_dictionary}
\title{Read OEPS data documentation}
\usage{
load_oeps_dictionary(scale)
}
\arguments{
\item{scale}{String specifying which geographic scale to pull variables for.
Must be one of "tract", "zcta", "county", or "state".}
}
\value{
A tibble containing the data dictionary.
}
\description{
\code{oeps_dictionary} returns the data dictionary for OEPS data at a given
geographic scale.
}
\examples{
tract_dictionary <- load_oeps_dictionary("tract")
View(tract_dictionary) # for optimal viewing experience

}
