% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables.R, R/wrangle_data.R
\name{merge_keys}
\alias{merge_keys}
\title{Merge keys}
\usage{
merge_keys()

merge_keys()
}
\value{
A list of strings (the names of the merge key columns).

A list of strings (the names of the merge key columns).
}
\description{
Returns the list of merge keys used by all data within this package.

Returns the list of merge keys used by all data within this package.
}
