% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeps_dictionary.R
\name{oeps_dictionary}
\alias{oeps_dictionary}
\title{Preview OEPS data}
\usage{
oeps_dictionary(scale)
}
\arguments{
\item{scale}{String specifying which geographic scale to pull variables for.
Must be one of "tract", "zcta", "county", or "state".}
}
\value{
A tibble cotnaining the data dictionary.
}
\description{
\code{oeps_dictionary} returns the data dictionary for OEPS data at a given
geographic scale.
}
\examples{
tract_dictionary <- oeps_dictionary('tract')
View(tract_dictionary) # for optimal viewing experience

}
