% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{retrieve_geometry}
\alias{retrieve_geometry}
\title{Retrieve OEPS geometries}
\usage{
retrieve_geometry(scale, quiet = FALSE, cache = TRUE)
}
\arguments{
\item{scale}{String specifying geographic scale to pull from.}

\item{quiet}{Boolean specifying whether to output st_read console output.}

\item{cache}{Boolean specifying whether to use the cache. If t}
}
\value{
Simple feature object containing the specified geometry.
}
\description{
Either download OEPS geometries, or use cached versions.
}
