% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tidify_data}
\alias{tidify_data}
\title{Tidify non-merge key variables}
\usage{
tidify_data(df)
}
\arguments{
\item{df}{Dataframe to convert to wide format.}
}
\value{
Dataframe in tidy format.
}
\description{
Convert dataframe from wide to tidy format, ignoring merge keys specified
in \code{merge_keys}.
}
\seealso{
\code{\link[=merge_keys]{merge_keys()}}
}
