% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{translate_to_fips}
\alias{translate_to_fips}
\title{Translate to FIPS}
\usage{
translate_to_fips(state, county = NULL)
}
\arguments{
\item{state}{Name, abbreviation, or FIPS of the state to be converted}

\item{county}{Name or abbreviation of the county to be converted. Optional}
}
\value{
FIPS code for the state.
}
\description{
Take State and possibly County names and return the FIPS code of the relevant
unit.
}
