% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_formatting.R
\name{county_to_fips}
\alias{county_to_fips}
\title{Identify counties GEOID}
\usage{
county_to_fips(county, state_fips)
}
\arguments{
\item{county}{As in \link{filter_by_geography}}

\item{state_fips}{The FIPS code for one of the 50 states or DC.}
}
\value{
Single County GEOID or list of GEOIDS.
}
\description{
Takes one or more county names, alongside their relevant
state, and return the counties' GEOIDS.
}
