% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_oeps.R
\name{load_oeps}
\alias{load_oeps}
\title{Load OEPS data}
\usage{
load_oeps(
  scale,
  year,
  themes = "All",
  states = NULL,
  counties = NULL,
  tidy = FALSE,
  geometry = FALSE,
  cache = TRUE
)
}
\arguments{
\item{scale}{String specifying which geographic scale to pull variables for.
Must be one of "tract", "zcta", "county", or "state".}

\item{year}{String or integer specifying which year to pull variables for.
Must be one of "1980", "1990", "2000", "2010", or "Latest".}

\item{themes}{String or list of strings specifying which Opioid Risk Environment
themes to pull variables for. Valid themes are "Geography", "Social", "Environment",
"Economic", "Outcome", "Policy", "Composite", or "All". Defaults to "All".}

\item{states}{String or vector of strings specifying which states to pull data for.
Should be FIPS codes. Cannot be a vector if counties is provided.}

\item{counties}{String or vector of strings specifying which counties to pull data for.
Should be in FIPS codes, and is invalid if scale is states, or if no or multiple
states are specified.}

\item{tidy}{Boolean specifying whether to return data in tidy format. Defaults False.}

\item{geometry}{Boolean specifying whether to pull geometries for the dataset
TRUE or not FALSE. Defaults FALSE.}

\item{cache}{Boolean specifying whether to use/create cached geometries and
attribute data (TRUE) or to pull fresh geometry and attribute data (FALSE).
To update cached geometries and data, call \code{\link[=clear_cache]{clear_cache()}} followed by
one of \code{\link[=cache_geometries]{cache_geometries()}}, or \code{\link[=cache_oeps_tables]{cache_oeps_tables()}}.}
}
\value{
If geometry is FALSE, a tibble containing the requested variables.
If geometry is TRUE, returns a simple feature collection.
}
\description{
\code{load_oeps} packages opioid environment policy scan data with given geographic
scale, year, and themes. Can also pull geometries.
}
\examples{
tracts_2010 <- load_oeps("tract", 2010, geometry = TRUE)
county_latest <- load_oeps("county", "latest", themes = "social")

# themes can also be a list of options
zcta_latest <- load_oeps("zcta", "latest", c("composite", "economic"))

}
\seealso{
\code{\link[=load_oeps_dictionary]{load_oeps_dictionary()}}, which loads a basic data dictionary
showing viable datasets.
}
