% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_formatting.R
\name{merge_keys}
\alias{merge_keys}
\title{Merge keys}
\usage{
merge_keys()
}
\value{
A vector of strings (the names of the merge key columns).
}
\description{
Returns the vector of merge keys used by all data within this package.
}
