% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{retrieve_geometry}
\alias{retrieve_geometry}
\title{Retrieve OEPS geometries}
\usage{
retrieve_geometry(scale, quiet = FALSE)
}
\arguments{
\item{scale}{String specifying geographic scale to pull from.
One of 'state', 'tract', 'county', or 'zcta'}

\item{quiet}{Boolean specifying whether to output st_read console output.
Defaults to FALSE.}
}
\value{
Simple feature collection containing the specified geometries.
}
\description{
Download OEPS geometries of a specified scale.
}
