% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables.R
\name{standardize_scale}
\alias{standardize_scale}
\title{Standardize scale}
\usage{
standardize_scale(scale)
}
\description{
Take user defined scale and try to map it to one of a selection of possible
inputs.
}
