% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_formatting.R
\name{state_to_fips}
\alias{state_to_fips}
\title{Identify a state's FIPS code.}
\usage{
state_to_fips(state)
}
\arguments{
\item{state}{String or vector of strings detailing which states to filter
on. Can be the state's name, abbreviation, or GEOID.}
}
\description{
Takes a states name, abbreviation, or FIPS code and returns the
FIPS code.
}
